module_adapter: modules: Change module entry point type to uintptr_t
Change type of the entry_point parameter in the native_system_agent_start
and system_agent_start functions to uintptr_t which allows safe pointer
passing.

+++ src/audio/module_adapter/iadk/system_agent.cpp
@@ -124,7 +124,7 @@ int SystemAgent::CheckIn(ProcessingModuleFactoryInterface& module_factory,
 typedef int (*create_instance_f)(uint32_t module_id, uint32_t instance_id, uint32_t core_id,
 				 void *mod_cfg, void *parent_ppl, void **mod_ptr);
 
-int system_agent_start(uint32_t entry_point, uint32_t module_id, uint32_t instance_id,
+int system_agent_start(uintptr_t entry_point, uint32_t module_id, uint32_t instance_id,
 		       uint32_t core_id, uint32_t log_handle, void* mod_cfg,
 		       void **adapter)
 {
